Bootstrap.Generator -> Framework.Generator
