+++ src/Generators/Elegy.Framework.Generator/Elegy.Framework.Generator.csproj
